% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citygeocoder-package.r
\docType{package}
\name{citygeocoder}
\alias{citygeocoder}
\title{Standard and reverse geocoding for US cities/populated places.}
\description{
Standard and reverse geocoding for US cities/populated places.
}
\author{
Caleb Jenkins (@cjtexas)
}
