% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{geocode}
\usage{
geocode(search_city = "Boerne", search_state = "TX")
}
\arguments{
\item{search_city}{character vector}

\item{search_state}{character vector}
}
\description{
find the longitude and latitude for cities/populated places in the US
}
