% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_geocode.R
\name{reverse_geocode}
\alias{reverse_geocode}
\title{reverse_geocode}
\usage{
reverse_geocode(lon = -101.68181, lat = 34.61681)
}
\arguments{
\item{lon}{numeric vector}

\item{lat}{numeric vector}
}
\description{
find the closest US city/populated place to any given longitude/latitude pair(s)
}
